Added software locking in addition to hardware locking for PHY access.
The 82571 controller is a two port Ethernet card. When accessing PHY
registers, one must not only obtain a lock to prevent collisions with
hardware, but also a software lock to avoid any conflicts with other
drivers that might be controlling the other Ethernet port. For
example, I have one port bound to Snabb Switch, while the other port
in the card is bound to the e1000e driver to the OS.